requirements.txt
torch>=2.0.0
torchvision>=0.15.0
facenet-pytorch>=2.5.3
scikit-learn>=1.3.0
opencv-python>=4.8.0
flask>=2.0.0
werkzeug>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
flask-wtf>=1.1.0
email-validator>=2.0.0
pyopenssl>=23.0.0